update：修改代码

+++ csagent-modules/csagent-knowledge/src/main/java/com/csagent/knowledge/domain/KbQuestion.java
@@ -40,7 +40,7 @@ public class KbQuestion extends TenantEntity {
     /**
      * 命中次数
      */
-    private Long hitNum;
+    private Integer hitNum;
 
     /**
      * 修改部门
